pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "helmgate"
version = "0.1.0"
description = "Helm chart linter and policy enforcement CLI"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12",
    "pyyaml>=6.0",
    "rich>=13.0",
    "cryptography>=42.0",
]

[project.scripts]
helmgate = "helmgate.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["helmgate*"]
